reorganized reducers
