--- src/auth/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import auth from './auth';
-import message from './message';
-
-export default combineReducers({
-  auth,
-  message,
-});
