+++ src/store/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import auth from '../auth/reducers/auth';
+import message from '../auth/reducers/message';
+import taskReducer from '../tasks/reducer';
+
+export default combineReducers({
+  auth,
+  message,
+  taskReducer,
+});
