+++ src/store1.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
-import rootReducer from './auth/reducers';
+import rootReducer from './store/reducers';
 
 const middleware = [thunk];
 const store = createStore(
